+++ src/geometry/ShellSceneNode.cxx
@@ -26,7 +26,7 @@
 // FIXME (SceneRenderer.cxx is in src/bzflag)
 #include "SceneRenderer.h"
 
-#define	ShellRadius1_2	(M_SQRT1_2 * ShellRadius)
+#define	ShellRadius1_2	((const GLfloat)(M_SQRT1_2 * ShellRadius))
 
 const GLfloat		ShellSceneNode::shellVertex[9][3] = {
 				{ 3.0f * ShellRadius, 0.0f, 0.0f },
@@ -42,13 +42,13 @@ const GLfloat		ShellSceneNode::shellVertex[9][3] = {
 const GLfloat		ShellSceneNode::shellNormal[10][3] = {
 				{ 1.0f, 0.0f, 0.0f },
 				{ 0.0f, -1.0f, 0.0f },
-				{ 0.0f, -M_SQRT1_2, -M_SQRT1_2 },
+				{ 0.0f, (float)-M_SQRT1_2, (float)-M_SQRT1_2 },
 				{ 0.0f, 0.0f, -1.0f },
-				{ 0.0f, M_SQRT1_2, -M_SQRT1_2 },
+				{ 0.0f, (float)M_SQRT1_2, (float)-M_SQRT1_2 },
 				{ 0.0f, 1.0f, 0.0f },
-				{ 0.0f, M_SQRT1_2, M_SQRT1_2 },
+				{ 0.0f, (float)M_SQRT1_2, (float)M_SQRT1_2 },
 				{ 0.0f, 0.0f, 1.0f },
-				{ 0.0f, -M_SQRT1_2, M_SQRT1_2 },
+				{ 0.0f, (float)-M_SQRT1_2, (float)M_SQRT1_2 },
 				{-1.0f, 0.0f, 0.0f }
 			};
 
@@ -76,8 +76,8 @@ void			ShellSceneNode::move(const GLfloat pos[3],
 						const GLfloat forward[3])
 {
   setCenter(pos);
-  azimuth = 180.0f / M_PI*atan2f(forward[1], forward[0]);
-  elevation = -180.0f / M_PI*atan2f(forward[2], hypotf(forward[0],forward[1]));
+  azimuth = (float)(180.0 / M_PI*atan2f(forward[1], forward[0]));
+  elevation = (float)(-180.0 / M_PI*atan2f(forward[2], hypotf(forward[0],forward[1])));
 }
 
 void			ShellSceneNode::notifyStyleChange()
